Extract context policy helpers into _policy
Move the allowlist, hook-merge, and timeout narrowing rules out of
cuprum/context/core.py into a dedicated cuprum/context/_policy.py so the
invariants can be tested as pure functions. Unify the three per-hook merge
helpers into a single _merge_hooks with an explicit ordering flag, and split
the oversized test_context module into narrowing, timeout, and isolation
suites.
