+++ cuprum/context/__init__.py
@@ -44,6 +44,15 @@ construction.
 
 """
 
+from cuprum.context._policy import (
+    _narrow_allowlist as _narrow_allowlist,
+)
+from cuprum.context._policy import (
+    _resolve_narrowed_timeout as _resolve_narrowed_timeout,
+)
+from cuprum.context._policy import (
+    _validate_timeout as _validate_timeout,
+)
 from cuprum.context.core import (
     AfterHook,
     BeforeHook,
@@ -51,24 +60,6 @@ from cuprum.context.core import (
     CuprumContext,
     ForbiddenProgramError,
     ScopeConfig,
-)
-from cuprum.context.core import (
-    _merge_after_hooks as _merge_after_hooks,
-)
-from cuprum.context.core import (
-    _merge_before_hooks as _merge_before_hooks,
-)
-from cuprum.context.core import (
-    _merge_observe_hooks as _merge_observe_hooks,
-)
-from cuprum.context.core import (
-    _narrow_allowlist as _narrow_allowlist,
-)
-from cuprum.context.core import (
-    _resolve_narrowed_timeout as _resolve_narrowed_timeout,
-)
-from cuprum.context.core import (
-    _validate_timeout as _validate_timeout,
 )
 from cuprum.context.env_overlay import merge_env_overlays, resolve_env
 from cuprum.context.registration import (
